refactor: move sync_wallet method from api.rs to sync.rs
The sync_wallet method logically belongs with other sync-related
functionality rather than in the general API module.
